// package/src/app/api/user/models/User.ts
export interface User {
  id: number;
  firstName: string;
  lastName: string;
  middleName: string;
  dateOfBirth: string;
  branch: string;
  plateNumber: string;
  emergencyFullName: string;
  emergencyPhone: string;
  emergencyAddress: string;
  unit: string;
  unionName: string;
  ninBvn: string;
  address: string;
  image: string;
  printed: string;
  abksthNo: string;
}